jupyter/requirements.txt
kafka-python
FlightRadarAPI
geopy
